Add file check to clearPRIdQueue.

+++ gha_cron.php
@@ -196,7 +196,9 @@ function savePRIdsQueue($PRList) {
 
 function clearPRIdsQueue() {
     global $PRList_file;
-    unlink($PRList_file);
+    if ( file_exists($PRList_file) ) {
+        unlink($PRList_file);
+    }
 }
